requirements (2).txt
streamlit>=1.28.0
psycopg2-binary>=2.9.7
bcrypt>=4.1.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
scikit-learn>=1.3.0
nltk>=3.8.0
textblob>=0.17.1
requests>=2.31.0
python-multipart>=0.0.6